projects/1_JobAutoApply/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
selenium
python-dotenv
langdetect
scikit-learn
nltk==3.8.1
spacy==3.7.2
langchain
openai
pinecone
scikit-learn
numpy
langchain-community
langchain-openai
anthropic